TIP-860: Move Locale related api tests
