+++ tests/back/Channel/EndToEnd/Locale/ExternalApi/FilterLocaleEndToEnd.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\Locale;
+namespace AkeneoTest\Channel\EndToEnd\Locale\ExternalApi;
 
-use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class FilterLocaleIntegration extends ApiTestCase
+class FilterLocaleEndToEnd extends ApiTestCase
 {
     public function testFilterActivatedLocales()
     {
